chore: add bincode compat
