+++ crates/primitives/src/alloy_compat.rs
@@ -7,92 +7,92 @@ use alloy_network::{AnyHeader, AnyRpcBlock, AnyRpcTransaction, AnyTxEnvelope};
 use alloy_serde::WithOtherFields;
 use op_alloy_rpc_types as _;
 
-impl TryFrom<AnyRpcBlock> for Block {
-    type Error = alloy_rpc_types::ConversionError;
-
-    fn try_from(block: AnyRpcBlock) -> Result<Self, Self::Error> {
-        use alloy_rpc_types::ConversionError;
-
-        let block = block.inner;
-
-        let transactions = {
-            let transactions: Result<Vec<TransactionSigned>, ConversionError> = match block
-                .transactions
-            {
-                alloy_rpc_types::BlockTransactions::Full(transactions) => {
-                    transactions.into_iter().map(|tx| tx.try_into()).collect()
-                }
-                alloy_rpc_types::BlockTransactions::Hashes(_) |
-                alloy_rpc_types::BlockTransactions::Uncle => {
-                    // alloy deserializes empty blocks into `BlockTransactions::Hashes`, if the tx
-                    // root is the empty root then we can just return an empty vec.
-                    if block.header.transactions_root == EMPTY_TRANSACTIONS {
-                        Ok(Vec::new())
-                    } else {
-                        Err(ConversionError::Custom("missing transactions".to_string()))
-                    }
-                }
-            };
-            transactions?
-        };
-
-        let AnyHeader {
-            parent_hash,
-            ommers_hash,
-            beneficiary,
-            state_root,
-            transactions_root,
-            receipts_root,
-            logs_bloom,
-            difficulty,
-            number,
-            gas_limit,
-            gas_used,
-            timestamp,
-            extra_data,
-            mix_hash,
-            nonce,
-            base_fee_per_gas,
-            withdrawals_root,
-            blob_gas_used,
-            excess_blob_gas,
-            parent_beacon_block_root,
-            requests_hash,
-        } = block.header.inner;
-
-        Ok(Self {
-            header: Header {
-                parent_hash,
-                ommers_hash,
-                beneficiary,
-                state_root,
-                transactions_root,
-                receipts_root,
-                logs_bloom,
-                difficulty,
-                number,
-                gas_limit,
-                gas_used,
-                timestamp,
-                extra_data,
-                mix_hash: mix_hash
-                    .ok_or_else(|| ConversionError::Custom("missing mixHash".to_string()))?,
-                nonce: nonce.ok_or_else(|| ConversionError::Custom("missing nonce".to_string()))?,
-                base_fee_per_gas,
-                withdrawals_root,
-                blob_gas_used,
-                excess_blob_gas,
-                parent_beacon_block_root,
-                requests_hash,
-            },
-            body: BlockBody {
-                transactions,
-                ommers: Default::default(),
-                withdrawals: block.withdrawals.map(|w| w.into_inner().into()),
-            },
-        })
-    }
-}
+// impl TryFrom<AnyRpcBlock> for Block {
+//     type Error = alloy_rpc_types::ConversionError;
+//
+//     fn try_from(block: AnyRpcBlock) -> Result<Self, Self::Error> {
+//         use alloy_rpc_types::ConversionError;
+//
+//         let block = block.inner;
+//
+//         let transactions = {
+//             let transactions: Result<Vec<TransactionSigned>, ConversionError> = match block
+//                 .transactions
+//             {
+//                 alloy_rpc_types::BlockTransactions::Full(transactions) => {
+//                     transactions.into_iter().map(|tx| tx.try_into()).collect()
+//                 }
+//                 alloy_rpc_types::BlockTransactions::Hashes(_) |
+//                 alloy_rpc_types::BlockTransactions::Uncle => {
+//                     // alloy deserializes empty blocks into `BlockTransactions::Hashes`, if the
+// tx                     // root is the empty root then we can just return an empty vec.
+//                     if block.header.transactions_root == EMPTY_TRANSACTIONS {
+//                         Ok(Vec::new())
+//                     } else {
+//                         Err(ConversionError::Custom("missing transactions".to_string()))
+//                     }
+//                 }
+//             };
+//             transactions?
+//         };
+//
+//         let AnyHeader {
+//             parent_hash,
+//             ommers_hash,
+//             beneficiary,
+//             state_root,
+//             transactions_root,
+//             receipts_root,
+//             logs_bloom,
+//             difficulty,
+//             number,
+//             gas_limit,
+//             gas_used,
+//             timestamp,
+//             extra_data,
+//             mix_hash,
+//             nonce,
+//             base_fee_per_gas,
+//             withdrawals_root,
+//             blob_gas_used,
+//             excess_blob_gas,
+//             parent_beacon_block_root,
+//             requests_hash,
+//         } = block.header.inner;
+//
+//         Ok(Self {
+//             header: Header {
+//                 parent_hash,
+//                 ommers_hash,
+//                 beneficiary,
+//                 state_root,
+//                 transactions_root,
+//                 receipts_root,
+//                 logs_bloom,
+//                 difficulty,
+//                 number,
+//                 gas_limit,
+//                 gas_used,
+//                 timestamp,
+//                 extra_data,
+//                 mix_hash: mix_hash
+//                     .ok_or_else(|| ConversionError::Custom("missing mixHash".to_string()))?,
+//                 nonce: nonce.ok_or_else(|| ConversionError::Custom("missing
+// nonce".to_string()))?,                 base_fee_per_gas,
+//                 withdrawals_root,
+//                 blob_gas_used,
+//                 excess_blob_gas,
+//                 parent_beacon_block_root,
+//                 requests_hash,
+//             },
+//             body: BlockBody {
+//                 transactions,
+//                 ommers: Default::default(),
+//                 withdrawals: block.withdrawals.map(|w| w.into_inner().into()),
+//             },
+//         })
+//     }
+// }
 
 impl TryFrom<AnyRpcTransaction> for TransactionSigned {
     type Error = alloy_rpc_types::ConversionError;
